requirements/base.txt
# Base requirements for all environments
flask==2.3.3
flask-restx==1.3.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-migrate==4.0.4
# Note: mysqlclient requires mysql development libraries and pkg-config
# For Docker builds, ensure these are installed in the Dockerfile
mysqlclient==2.2.0
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.3 